Commit nr 13 to the branch level-loader
Fixed the activate_spawners() method to actually change the  active flag in the spawners. Important note: to do so, I had to make the method reference an external module holding the class instances - in this case: config.Object_list. This is because the actual class instance we want to manipulate lives as a member of the Sprite group in config.Object_list. Changing the self.spawners active flag, even though intuitively correct, changes nothing in the instance living in the config module. This will likely translate to behaviors of other classes and methods as well, especially the deactivate method for spawners.